+++ odmlui/navigation_bar.py
@@ -1,3 +1,10 @@
+"""
+'navigation_bar' contains the 'NavigationBar' class.
+
+This class provides functionality to browse from
+a currently selected model object to the document root.
+"""
+
 import pygtkcompat
 
 import gtk
@@ -7,6 +14,15 @@ pygtkcompat.enable_gtk(version='3.0')
 
 
 class NavigationBar(gtk.Label):
+    """
+    The NavigationBar widget provides an interactive
+    link list from the Document root of the currently
+    active Document to the object that is currently
+    selected in the data model of the main application window.
+
+    Activating any of the links selects the respective object
+    and updates the application window accordingly.
+    """
     def __init__(self, *args, **kargs):
         super(NavigationBar, self).__init__(*args, **kargs)
         self._document = None
@@ -23,6 +39,9 @@ class NavigationBar(gtk.Label):
 
     @property
     def document(self):
+        """
+        :return: The currently active document.
+        """
         return self._document
 
     @document.setter
@@ -36,16 +55,30 @@ class NavigationBar(gtk.Label):
 
     @property
     def current_object(self):
+        """
+        :return: The currently selected object.
+        """
         return self._current_object
 
     @current_object.setter
     def current_object(self, obj):
+        """
+        Update the internal selected object with provided *obj*
+        and update the view accordingly.
+        """
         self._current_object = obj
         self.update_display()
         self.on_selection_change(obj)
 
     def switch(self, _, path):
-        """called if a link in the property_status Label widget is clicked"""
+        """
+        Retrieve the object corresponding to a provided path and
+        make it the selected one, updating the view. If no path
+        was provided use the NavigationBar's document as a
+        fallback object to update the view.
+
+        Called if a link in the NavigationBar's Label widget is clicked.
+        """
         if path:
             path = [int(i) for i in path.split(":")]
             obj = self._document.from_path(path)
@@ -56,8 +89,8 @@ class NavigationBar(gtk.Label):
 
     def set_model(self, obj):
         """
-        show the hierarchy for object *obj* and make
-        it the selected one
+        Show the hierarchy for object *obj* and make it
+        the selected one, updating the view.
         """
         self._current_hierarchy = [obj]
 
@@ -69,6 +102,11 @@ class NavigationBar(gtk.Label):
         self.current_object = cur
 
     def update_display(self):
+        """
+        Update the NavigationBar's label text with a
+        list of object links from the Document root
+        to the currently selected object.
+        """
         names = []
         cur = self._current_object
         for obj in self._current_hierarchy:
@@ -87,16 +125,18 @@ class NavigationBar(gtk.Label):
 
     def on_selection_change(self, obj):
         """
-        called whenever this widget elects a new current obj
+        Called whenever a new object *obj* in the underlying
+        data model is selected.
+
+        The actual method is set on the class at the point of usage.
         """
-        raise NotImplementedError
+        pass
 
     def on_section_changed(self, context):
         """
-        this is called by the Eventable modified MixIns of Value/Property/Section
-        and causes the GUI to refresh correspondingly
+        Refresh the NavigationBar display after the content
+        of the currently displayed document has been changed.
         """
-        # we are only interested in changes on sections
         if context.cur is not self._document:
             return